+++ vertx-bus/src/main/java/com/cloudentity/tools/vertx/conf/ConfService.java
@@ -10,4 +10,7 @@ public interface ConfService {
 
   @VertxEndpoint
   Future<JsonObject> getGlobalConf();
+
+  @VertxEndpoint
+  Future<JsonObject> getMaskedGlobalConf();
 }